Fixed inventory functions
Changed 'length' to 'len()' as it was previously invalid. Corrected the parameter for the functions. Changed the room item handling so that they are dictionary based instead of list based so that it works with the JSON. Corrected the undefined variables and missing quotes. Added return messages for AddItem, RemoveItem, PickUpItem and DropItem. Added a check to see if the item is in the inventory. Corrected the logic issues.